(Feature) 209: Alt text guidance
Uses css ::after pseudo-property to add guidance to alt text fields in
media attachment editors. This is the only viable way I could find to
add this text in - there are no PHP-based filters for amending the text,
as the media modals are all JS-based. I could not find a way to reliably
hook into the JS to add the text, so decided on css as the most
straightforward a reliable way of achieving what was needed.

Resolves:
https://trello.com/c/4ZuxMTwY/209-2-users-know-how-to-write-alt-text-for-images