(Feature) 283: Comment validation errors
Formats validation errors to match
http://govuk-elements.herokuapp.com/errors/. That guidance implies that
the validation should happen server-side, and the errors be returned on
page refresh, but WordPress' comment-handling doesn't really make that
possible.

So this is all JS-driven, but if you don't have JS, it'll fall back to
the default WP server-side validation. Aria attributes are used to make
the JS validation accessible.

Resolves:
https://trello.com/c/GKn8ICDr/283-standardise-validation-errors-on-comments-to-be-inline-with-gds-styles