(chore) Replace grunt-img with grunt-image
Grunt-img had not been updated in 5 years +, and contained a dependency
that resulted in a github security alert. Although it was only a dev
dependency, we are updating to reduce noise around security notices.

This commit replaces the package with grunt-image, and runs the new
`grunt image` task to optimise the image files. As per the previous
setup, this task does not move the image files, and does not run in the
standard `grunt` task.

Resolves: https://dxw.zendesk.com/agent/tickets/9875