Change all .visually-hidden classes to .visuallyhidden
The original card meant to fix an a11y problem on Firefox as the class .visually-hidden is almost identical to .visuallyhidden except the missing of `white-space: nowrap;`.

Although the recent releases of Firefox fixed the problem, we decided to rename all the .visually-hidden classes back to .visuallyhidden for consistency reason.

Resolves: https://trello.com/c/JqZMOm7U/98-fix-visually-hidden-css-class#comment-5df13230fee0bb132baa24b2