Unhook Roots' filter of stylesheet tags
The blogs theme includes functions from the Roots theme via composer.

One of these function filters style tags to remove media tags if the reference is `media="all"` or `media=""`.

It uses a regular expression to do so, and the format of stylesheet links output by WordPress in the admin no longer fits that regular expression. As a result, the function is simply breaking stylesheet links in the admin.

Ideally, this would be resolved by updating the version of Roots that the dxw/roots-composer package uses. However, we are now so far behind the current version (6.5.1 vs. 9.0.8), that to do so would be a big and risky undertaking. As this bug is blocking upgrades to WordPress 5.3.x (and hence security patches), we can't wait too long to fix this.

Therefore, the most straightforward approach looks like just unhooking that function, which is what I've done here.

Resolves: https://trello.com/c/Epfxqi6F/481-upgrade-blogs-wordpress-to-53