Always keep cookie banner hidden
The cookie banner no longer provides enough info/options to meet regulatory requirements.

On gov.uk blogs it will be replaced by a custom cookie control panel plugin.

However, because the cookie banner is hardcoded into the govuk template, it is not possible to remove it completely.

Therefore, I've chosen instead to always hide it in the CSS. This will also mean that if a site is using this theme as a parent theme, they will be able to re-show the banner via child theme CSS.

Resolves: https://trello.com/c/bf857VEO/485-identify-work-required-to-make-blogs-cookies-opt-in