Add govuk frontend in compatibility mode
This commit is the first step in migrating the theme to use the gov.uk Frontend Toolkit.

As per the instructions at https://deploy-preview-2--govuk-frontend-docs-prototype.netlify.com/migrating_from_legacy_products/#migrate-to-gov-uk-frontend-from-our-old-frameworks, the new frontend toolkit css, assets and JS have been included, with govuk-template compatibility mode activated.

No changes are currently visible in the frontend, which is the expected behaviour.

I have set the image & font assets paths as relative paths. This is because we are hoping to deploy both the WIP 3.0.0 theme and the master version of the theme to staging simultaneously. This will only work if the new version of the theme is at a path other than the standard wp-content/themes/gds-blogs path. So therefore an absolute path will not work.

Resolves: https://trello.com/c/4S3EMvh9/489-migrate-the-main-blog-theme-to-new-frontend-toolkit