Migrate index.php to govuk frontend
- Removes includes of partials that don't exist
- Updates the error message that would be displayed on a blog homepage if the blog has no posts, to bring it more inline with "There is a problem with the service" page pattern at https://design-system.service.gov.uk/patterns/problem-with-the-service-pages/ (though also using some of the error styling from https://design-system.service.gov.uk/components/error-summary/).

I'm not aware of a case where the error message part of this template has ever been served, but it could happen if either a) a blog was published with no posts on it; or b) there was some kind of database error when retrieving the posts. So the message should probably be sensible, just in case.

Resolves: https://trello.com/c/2AuPeJXp/500-template-to-migrate-indexphp