Move govuk template layout file out of build folder
We need to migrate the footer, header (and ultimately overall layout) to govuk frontend, so we don't want to be importing this as a dependency anymore, as we'll need to edit it and keep those edits.

So the first step is to copy it to somewhere that it won't get overwritten by the build process.

Resolves: https://trello.com/c/1sYDH7G9/505-migrate-header