Remove redundant sections in layout template
The cookie message div is no longer used, as this has been replaced by the civic cookie plugin. We were previously hiding it via CSS, but now we can edit the layout template we may as well remove it altogether.

I've also removed the global-header-bar div, as this was empty and had no effect on layout.

Resolves: https://trello.com/c/1sYDH7G9/505-migrate-header