Add an accessible version of the "Archives" widget
WordPress's default "Archives" widget submits the dropdown selection as soon as it is made. This can cause difficulties for users with accessibility requirements, and doesn't meet WCAG 2.0 AA.

This commit creates a new accessible version of that widget, with some default options (the archives are always monthly, the title is always "Archives"). It makes the component accessible by only submitting the selection (and causing the URL to change) once the accompanying "Go" button is hit.

The new widget is based on the model we used for the similar accessible Categories widget we created. As with the class, the tests really need a refactor, as they're too brittle at the moment (e.g. they fail if you add a space in an HTML component).

There's also an issue in that the new "Archives" component currently only works if you have JS activated - otherwise, the button does nothing. This will be remedied in a subsequent commit.

For: https://trello.com/c/tmxWq09B/24-jump-menu-on-archive-not-compliant