Show archives list in <ul> format if JS is disabled
The "Go" button in this archives widget only works if JS is enabled, so we need a non-JS alternative.

That alternative is to render the list of archive links as an HTML list. We show/hide the appropriate version using the `js-enabled` class on the body to toggle.

For: https://trello.com/c/tmxWq09B/24-jump-menu-on-archive-not-compliant