Escape co-author names in atom feeds
Before: while standard author names are escaped in atom feeds, co-author
names were not. This could result in the atom feed not serving valid xml

Now: we apply the esc_html filter to all author names in atom feeds.
esc_html is designed so as not to double-encode, so it's ok to call this
on standard author names that may have already been escaped.

Resolves: https://dxw.zendesk.com/agent/tickets/14996