Remove obselete jquery file
This commit removes the obselete version of jquery installed in the
theme, which means that the default installed version of jquery used
by wordpress will be used instead. Checking over the theme it doesn't
appear to be relying on javascript a great deal, and removing this
very old version resolves some error that were appearing in the console
while this file was present.