Separate import of govuk frontend js
Import statements is moved to a separate file as existing js in main.js cannot operate correctly when its type is defined as module. This is because its variables are needed by other js files on the site. The build process is adapted to copy the new file across to the build directory.