Restructure stylesheets, for consistency with GOV.UK Frontend Framework
Before, the theme's stylesheets were organised differently to those in the
GOV.UK Frontend Framework. There was misuse of folder names such as `common`,
which serve a different context in the framework.

Now, the structure and ordering of partials mirrors the framework. As a
reference, `govuk-frontend/dist/govuk/all.scss` and its imported files
helps to guide this work.

Although the files have been re-organised, there is still work to be
done to ensure that styles within those files are indeed in the correct
location. That's not the remit of this commit, though.