Relocate 'common' styles to relevant stylesheet partial
The styles originally contained in `core/common.scss` where a mix of
styles that applied to a range of different contexts e.g. global, components,
overrides. This commit moves all the styles to their relevant locations
and deletes the original stylesheet, in an attempt to bring a degree of
predictability to the codebase.

New stylesheet partials are added where appropriate.

Note that although the restructuring is opionated, the styles themeselves
have been left intact as that requries further investigation and
thought, as to whether further refactoring is beneficial, albeit very
likely.