Redirect non-atom feeds to /atom
Previously, other RSS feeds had been turned off, and it seemed that no
measures were put in place to manage the absence of these feeds, leading
to a 404 error.

However, a redirect has now been introduced for all non-Atom feeds,
providing users with an alternative.
This is a behaviour that happens in other of our sites/themes.

The reasons for their initial deactivation remain unclear to me.

The deactivation occurred in this PR:
https://github.com/dxw/govuk-blogs/pull/13