Enable selection of GOV.UK colour palette
This commit enables selection, within the Block Editor, of colours
from the Design System colour palette:

https://design-system.service.gov.uk/styles/colour/

For now, only colours that have a designated use are included e.g.
'black' is specified for 'text colour'. The only exceptions are, turquoise
is included as it seems to be the brand colour for blogs (evident in the
header border colour), and light grey is included as a neutral option
for component backgrounds.