Refactor theme.json palette to enable reuse of variables
Before, if a palette colour was to be reused in `theme.json` (e.g. to
constrain a palette for a block), the hex value of the colour would have
to be duplicated. This dilutes the source of truth, and increases the
chances of human error resulting in colours going out of sync.

Now, a group of custom presets is configured so that hex values only
need to be declared once, and then the generated CSS Custom Properties can
be referenced throughout `theme.json`.