Make spacing scale responsive
Before, the intention was to add custom CSS to accommodate the
responsive spacing sizes for small-viewports, from the GOV.UK spacing
scale:

https://design-system.service.gov.uk/styles/spacing/#responsive-spacing

However, CSS `clamp()` can be used to immitate media queries, and
therefore enable the responsive sizes to be added directly to
`theme.json`! The use of `1000` is an arbitrarily large number that
enables the `max` clamp value to kick-in at `40rem`.

See: https://css-tricks.com/responsive-layouts-fewer-media-queries/