Add govuk style buttons using block variations.
Using a variation on the core/buttons block and introducing style variations to the core/button block the experience of a custom gov.uk button block is created. This includes four style variations in line the govuk design system. Additional js adds an attribute to the front end to help with accessibility.

Add govuk style buttons using block variations.

Using a variation on the core/buttons block and introducing style variations to the core/button block the experience of a custom gov.uk button block is created. This includes four style variations in line the govuk design system. Additional js adds an attribute to the front end to help with accessibility.