Make navigation widget spacing consistent with other widgets
Before, the navigation widget had reduced vertical spacing if the main
navigation was selected as the source. This resulted in spacing that was
inconsistent with the other widgets in the sidebar.

Now, custom styling for the main nav widget has been removed, in favour
of adoption of the global sidebar styles.