Ensure accessible colour contrast in focus state of cookie banner buttons
Before, black text on dark green background would be initiated when the
button was focused, and combined with hover or active states. This
resulted in insufficent colour contrast for accessibility needs.

Now, the scope of the focus styles are constrained, to only apply if
neither the hover or active states are in play.