Rewrite non-existent authors to our selected archive author
Before: if a user had authored a post, but then the user had
subsequently been deleted without their posts being re-assigned, this
resulted in a blank author being displayed in the front-end, and the
"Author" spinner in the block editor never resolving.

Now: we can set an "archive_author" option in `wp_options`. This should
be set to the ID of the author we want to default to in cases where the
original author has been removed without re-assignment. When that option
is set, the next time the post is viewed in the front-end, the author
will be re-written to be the archive author.