Account for Co-Authors relationships when adjusting for deleted authors
Before: when we re-allocated the post author if the existing post author
no longer existed, we just changed the author id attached to the post.
This fixed the author display in the front-end, but did not resolve the
ever-spinning circle in the back-end, as if co-authors is active,
co-author relationships (defined by taxonomy term relationships) was
still in place.

Now: if we have amended the post author ID, we also remove any co-author
term relationships with the post in question. This means the newly
allocated post author now displays in both the front and back-ends.