Retain co-author data when setting archive author
Before: when we were setting the post author of a post who's post author
ID (according to the `author` value in the `wp_posts` table) no longer
existed, we also removed any co-authors attributed to that post

Now: we set the author ID in the `wp_posts` table to the archive author
ID, but the co-author attribution is retained. This ensures that e.g.
attribution for posts where the original user's WordPress user record
has been removed, but they still have a co-author profile, is maintained