Prompt editors to update captions in media modal
This change introduces a button in the media modal that has no
direct functionality but acts as a UI cue, encouraging editors to blur
the focused field, thereby triggering a caption update via
`wp_prepare_attachment_for_js`.

The solution adds a message field to display an "Update Caption" button
using custom HTML. While the button itself is non-functional, it
initiates a blur event to save changes in the media modal.

Additionally, the licence select field was reordered to prompt editors
to select the license last, ensuring fields are saved or updated
appropriately.